More AI
Tested the AStar pathfinding algorithm, finds the target node, takes the optimal path, But it seems like there are too many nodes
on the closed set by the end of the pathing.

The heuristic isn't perfect either.

Added factories for peas carrots and broccoli (including new controllers and entities that need to be programmed)
Also added factories for box characters and circle characters
Made spawn points for the enemies in the map and added them to the flag processor

Changed CharactorFactory to PlayerFactory.

Probably some other stuff, I forget.